// frontend/src/model/tripPricingInfoSeatType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */

export type TripPricingInfoSeatType = typeof TripPricingInfoSeatType[keyof typeof TripPricingInfoSeatType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TripPricingInfoSeatType = {
  NORMAL: 'NORMAL',
  VIP: 'VIP',
} as const;