// frontend/src/model/tripResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { RouteInfo } from './routeInfo';
import type { BusInfo } from './busInfo';
import type { TripResponseStatus } from './tripResponseStatus';
import type { TripPricingInfo } from './tripPricingInfo';

export interface TripResponse {
  id?: string;
  route?: RouteInfo;
  bus?: BusInfo;
  departureTime?: string;
  arrivalTime?: string;
  status?: TripResponseStatus;
  tripPricings?: TripPricingInfo[];
}